trt fixes 2
applied 26 change(s) touching 16 path(s)

changes:
  1. replace src/chatterbox/onnx_export/dynamic_shapes.py: Use graph-local symbolic dimension names to avoid ONNX exporter axis-name reuse warnings.
  2. replace src/chatterbox/models/speaker/exportable.py: Remove the unused speaker length binding from the export wrapper.
  3. replace src/chatterbox/onnx_export/graphs/speaker_encoder.py: Export the speaker encoder with only the fbank input to avoid an unused TensorRT binding.
  4. update src/chatterbox/onnx_export/graphs/s3_tokenizer.py: Use int32 tokenizer length inputs and int32 tokenizer outputs for TensorRT-friendly bindings.
  5. update src/chatterbox/onnx_export/graphs/s3_tokenizer.py: Record tokenizer integer graph bindings as int32.
  6. update src/chatterbox/onnx_export/graphs/reference_mel.py: Emit reference mel lengths as int32 to avoid TensorRT int64 output bindings.
  7. update src/chatterbox/onnx_export/graphs/reference_mel.py: Record reference mel length output dtype as int32.
  8. update src/chatterbox/onnx_export/graphs/token_to_mu.py: Use int32 token and length dummy inputs for token-to-mu export.
  9. update src/chatterbox/onnx_export/graphs/token_to_mu.py: Record token-to-mu integer graph bindings as int32.
  10. update src/chatterbox/onnx_export/modules/reference_mel.py: Keep the legacy reference mel export helper aligned with int32 length outputs.
  11. update src/chatterbox/onnx_export/modules/s3_tokenizer.py: Keep the legacy tokenizer export helper aligned with int32 length inputs.
  12. replace src/chatterbox/onnx_export/modules/speaker_encoder.py: Keep the legacy speaker helper aligned with the single-input speaker graph.
  13. update src/chatterbox/onnx_export/modules/token_to_mu.py: Keep the legacy token-to-mu helper aligned with int32 token and length bindings.
  14. update src/chatterbox/tensorrt/shapes.py: Keep token profile maxima consistent with the exported positional-encoding limit.
  15. update src/chatterbox/tensorrt/shapes.py: Remove the unused speaker length input from TensorRT default profiles.
  16. update src/chatterbox/tensorrt/shapes.py: Reject token-to-mu TensorRT profiles whose independent maxima exceed the supported total token budget.
  17. update src/chatterbox/tensorrt/cuda.py: Expose CUDA memory availability for safer TensorRT workspace sizing.
  18. update src/chatterbox/tensorrt/builder.py: Cap TensorRT workspace against available CUDA memory to avoid oversized tactic warnings.
  19. update src/chatterbox/tensorrt/builder.py: Add a bounded effective workspace calculation for TensorRT builds.
  20. update src/chatterbox/tensorrt/builder.py: Apply the effective workspace cap when configuring TensorRT builds.
  21. update src/chatterbox/tensorrt/config.py: Use a conservative default TensorRT workspace size.
  22. update src/chatterbox/tensorrt/cli.py: Align the TensorRT CLI workspace default with the conservative builder default.
  23. update src/chatterbox/onnx_export/README.md: Document int32 tokenizer bindings in the ONNX graph signature.
  24. update src/chatterbox/onnx_export/README.md: Document the single-input speaker encoder graph signature.
  25. update src/chatterbox/onnx_export/README.md: Document int32 reference mel length output.
  26. update src/chatterbox/onnx_export/README.md: Document int32 token-to-mu integer bindings.

updated:
  src/chatterbox/onnx_export/README.md
  src/chatterbox/onnx_export/graphs/reference_mel.py
  src/chatterbox/onnx_export/graphs/s3_tokenizer.py
  src/chatterbox/onnx_export/graphs/token_to_mu.py
  src/chatterbox/onnx_export/modules/reference_mel.py
  src/chatterbox/onnx_export/modules/s3_tokenizer.py
  src/chatterbox/onnx_export/modules/token_to_mu.py
  src/chatterbox/tensorrt/builder.py
  src/chatterbox/tensorrt/cli.py
  src/chatterbox/tensorrt/config.py
  src/chatterbox/tensorrt/cuda.py
  src/chatterbox/tensorrt/shapes.py

replaced:
  src/chatterbox/models/speaker/exportable.py
  src/chatterbox/onnx_export/dynamic_shapes.py
  src/chatterbox/onnx_export/graphs/speaker_encoder.py
  src/chatterbox/onnx_export/modules/speaker_encoder.py

+++ src/chatterbox/onnx_export/graphs/reference_mel.py
@@ -90,7 +90,7 @@ class ReferenceMel24kExport(torch.nn.Module):
         prompt_feat_len = torch.full(
             (prompt_feat.size(0),),
             prompt_feat.size(1),
-            dtype=torch.long,
+            dtype=torch.int32,
             device=prompt_feat.device,
         )
         return prompt_feat, prompt_feat_len
@@ -113,5 +113,5 @@ REFERENCE_MEL_24K_SPEC = GraphSpec(
     make_module=make_module,
     make_dummy_inputs=make_dummy_inputs,
     input_dtypes={"wav_24k": "float32"},
-    output_dtypes={"prompt_feat": "float32", "prompt_feat_len": "int64"},
+    output_dtypes={"prompt_feat": "float32", "prompt_feat_len": "int32"},
 )
